Separate projects from home page, also fix errant unit test
