--- ClientApp/src/app/home-page/home-page.component.css
@@ -1,14 +1,0 @@
-#project-cards-container {
-  width: 100%;
-
-  display: grid;
-  grid-template-columns: repeat(2, 1fr);
-  column-gap: var(--content-padding);
-  row-gap: var(--content-padding);
-}
-
-@media only screen and (max-width: 768px) {
-  #project-cards-container {
-    grid-template-columns: repeat(1, auto);
-  }
-}
