+++ ClientApp/src/app/home-page/home-page.component.ts
@@ -1,27 +1,15 @@
-import {
-  Component,
-  OnInit,
-} from '@angular/core';
-
-import { ServerApi } from '../server-api/server-api.service';
-
-import { ProjectData } from '../../types/project';
+import { Component, OnInit } from '@angular/core';
 
 @Component({
   selector: 'home-page',
   templateUrl: './home-page.component.html',
-  styleUrls: ['./home-page.component.css'],
+  styleUrls: ['./home-page.component.css']
 })
 export class HomePageComponent implements OnInit {
-  projects!: ProjectData[];
 
-  constructor(private serverApi: ServerApi) { }
+  constructor() { }
 
   ngOnInit(): void {
-    this.fetchProjectData();
   }
 
-  fetchProjectData() {
-    this.projects = this.serverApi.getAllProjects();
-  }
 }
