+++ ClientApp/src/app/project-page/project-page.component.spec.ts
@@ -87,12 +87,13 @@ describe('ProjectPageComponent', () => {
   });
 
   it('should order sprints correctly', () => {
+    const blankDate = new Date();
     const test1: SprintData[] = [
       {
         id: 0,
         name: '',
-        startDate: new Date(),
-        endDate: new Date(),
+        startDate: blankDate,
+        endDate: blankDate,
         isCompleted: false,
         pointsAttempted: 0,
         pointsCompleted: 0,
@@ -103,8 +104,8 @@ describe('ProjectPageComponent', () => {
       {
         id: 1,
         name: '',
-        startDate: new Date(),
-        endDate: new Date(),
+        startDate: blankDate,
+        endDate: blankDate,
         isCompleted: false,
         pointsAttempted: 0,
         pointsCompleted: 0,
@@ -117,8 +118,8 @@ describe('ProjectPageComponent', () => {
       {
         id: 1,
         name: '',
-        startDate: new Date(),
-        endDate: new Date(),
+        startDate: blankDate,
+        endDate: blankDate,
         isCompleted: false,
         pointsAttempted: 0,
         pointsCompleted: 0,
@@ -129,8 +130,8 @@ describe('ProjectPageComponent', () => {
       {
         id: 0,
         name: '',
-        startDate: new Date(),
-        endDate: new Date(),
+        startDate: blankDate,
+        endDate: blankDate,
         isCompleted: false,
         pointsAttempted: 0,
         pointsCompleted: 0,
@@ -147,8 +148,8 @@ describe('ProjectPageComponent', () => {
       {
         id: 0,
         name: '',
-        startDate: new Date(),
-        endDate: new Date(),
+        startDate: blankDate,
+        endDate: blankDate,
         isCompleted: false,
         pointsAttempted: 0,
         pointsCompleted: 0,
@@ -159,8 +160,8 @@ describe('ProjectPageComponent', () => {
       {
         id: 1,
         name: '',
-        startDate: new Date(),
-        endDate: new Date(),
+        startDate: blankDate,
+        endDate: blankDate,
         isCompleted: true,
         pointsAttempted: 0,
         pointsCompleted: 0,
@@ -173,8 +174,8 @@ describe('ProjectPageComponent', () => {
       {
         id: 1,
         name: '',
-        startDate: new Date(),
-        endDate: new Date(),
+        startDate: blankDate,
+        endDate: blankDate,
         isCompleted: true,
         pointsAttempted: 0,
         pointsCompleted: 0,
@@ -185,8 +186,8 @@ describe('ProjectPageComponent', () => {
       {
         id: 0,
         name: '',
-        startDate: new Date(),
-        endDate: new Date(),
+        startDate: blankDate,
+        endDate: blankDate,
         isCompleted: false,
         pointsAttempted: 0,
         pointsCompleted: 0,
@@ -204,7 +205,7 @@ describe('ProjectPageComponent', () => {
         id: 0,
         name: '',
         startDate: new Date('19 Jan 2023 00:00:00 GMT'),
-        endDate: new Date(),
+        endDate: blankDate,
         isCompleted: false,
         pointsAttempted: 0,
         pointsCompleted: 0,
@@ -216,7 +217,7 @@ describe('ProjectPageComponent', () => {
         id: 1,
         name: '',
         startDate: new Date('18 Jan 2023 00:00:00 GMT'),
-        endDate: new Date(),
+        endDate: blankDate,
         isCompleted: false,
         pointsAttempted: 0,
         pointsCompleted: 0,
@@ -230,7 +231,7 @@ describe('ProjectPageComponent', () => {
         id: 1,
         name: '',
         startDate: new Date('18 Jan 2023 00:00:00 GMT'),
-        endDate: new Date(),
+        endDate: blankDate,
         isCompleted: false,
         pointsAttempted: 0,
         pointsCompleted: 0,
@@ -242,7 +243,7 @@ describe('ProjectPageComponent', () => {
         id: 0,
         name: '',
         startDate: new Date('19 Jan 2023 00:00:00 GMT'),
-        endDate: new Date(),
+        endDate: blankDate,
         isCompleted: false,
         pointsAttempted: 0,
         pointsCompleted: 0,
