+++ ClientApp/src/app/projects-page/projects-page.component.css
@@ -1,0 +1,14 @@
+#project-cards-container {
+  width: 100%;
+
+  display: grid;
+  grid-template-columns: repeat(2, 1fr);
+  column-gap: var(--content-padding);
+  row-gap: var(--content-padding);
+}
+
+@media only screen and (max-width: 768px) {
+  #project-cards-container {
+    grid-template-columns: repeat(1, auto);
+  }
+}
